fix: fixed all the comments
