+++ src/pages/Home.tsx
@@ -1,12 +1,13 @@
 import { Map } from 'components/Map';
 import { SearchBar } from 'components/Header';
+import styled from 'styled-components';
 
 const Home: React.FC = () => {
   return (
-    <div style={{ height: '100%', width: '100%' }}>
+    <>
       <SearchBar />
       <Map />
-    </div>
+    </>
   );
 };
 export default Home;
